Move `eval_map` to mmrotate.evaluation.
