+++ mmrotate/core/evaluation/__init__.py
@@ -1,0 +1,5 @@
+# Copyright (c) OpenMMLab. All rights reserved.
+from .eval_map import eval_map
+from .recall import eval_recalls
+
+__all__ = ['eval_map', 'eval_recalls']
